+++ clients/derivativestradingcoinfutures/src/restapi/models/model_query_current_open_order_response.go
@@ -39,7 +39,7 @@ type QueryCurrentOpenOrderResponse struct {
 	Side *string `json:"side,omitempty"`
 	// Position side
 	PositionSide *string `json:"positionSide,omitempty"`
-	// Enum：completed，processing
+	// Status.
 	Status *string `json:"status,omitempty"`
 	// please ignore when order type is TRAILING_STOP_MARKET
 	StopPrice *string `json:"stopPrice,omitempty"`