+++ clients/derivativestradingcoinfutures/src/websocketapi/models/model_query_order_response.go
@@ -18,8 +18,7 @@ var _ common.MappedNullable = &QueryOrderResponse{}
 // QueryOrderResponse struct for QueryOrderResponse
 type QueryOrderResponse struct {
 	// positionId
-	Id *string `json:"id,omitempty"`
-	// Enum：completed，processing
+	Id     *string                   `json:"id,omitempty"`
 	Status *int64                    `json:"status,omitempty"`
 	Result *QueryOrderResponseResult `json:"result,omitempty"`
 	// Rate limit definitions.