+++ clients/derivativestradingportfoliomargin/src/restapi/models/model_cancel_margin_account_order_response.go
@@ -33,7 +33,7 @@ type CancelMarginAccountOrderResponse struct {
 	ExecutedQty *string `json:"executedQty,omitempty"`
 	// Cummulative Quote Qty.
 	CummulativeQuoteQty *string `json:"cummulativeQuoteQty,omitempty"`
-	// Enum：completed，processing
+	// Status.
 	Status *string `json:"status,omitempty"`
 	// Time In Force.
 	TimeInForce *string `json:"timeInForce,omitempty"`