+++ clients/derivativestradingportfoliomargin/src/restapi/models/model_margin_account_new_oco_response_order_reports_inner.go
@@ -35,7 +35,7 @@ type MarginAccountNewOcoResponseOrderReportsInner struct {
 	ExecutedQty *string `json:"executedQty,omitempty"`
 	// Cummulative Quote Qty.
 	CummulativeQuoteQty *string `json:"cummulativeQuoteQty,omitempty"`
-	// Enum：completed，processing
+	// Status.
 	Status *string `json:"status,omitempty"`
 	// Time In Force.
 	TimeInForce *string `json:"timeInForce,omitempty"`