+++ clients/derivativestradingportfoliomargin/src/restapi/models/model_query_all_cm_orders_response_inner.go
@@ -39,7 +39,7 @@ type QueryAllCmOrdersResponseInner struct {
 	Side *string `json:"side,omitempty"`
 	// BOTH means that it is the position of One-way Mode
 	PositionSide *string `json:"positionSide,omitempty"`
-	// Enum：completed，processing
+	// Status.
 	Status *string `json:"status,omitempty"`
 	// Trade symbol, if existing.
 	Symbol *string `json:"symbol,omitempty"`