+++ clients/derivativestradingportfoliomargin/src/restapi/models/model_query_users_cm_force_orders_response_inner.go
@@ -23,7 +23,7 @@ type QueryUsersCmForceOrdersResponseInner struct {
 	Symbol *string `json:"symbol,omitempty"`
 	// Pair.
 	Pair *string `json:"pair,omitempty"`
-	// Enum：completed，processing
+	// Status.
 	Status *string `json:"status,omitempty"`
 	// Client Order ID.
 	ClientOrderId *string `json:"clientOrderId,omitempty"`