+++ clients/derivativestradingusdsfutures/tests/unit/websocketstreams/api_public_test.go
@@ -119,7 +119,7 @@ func Test_binancederivativestradingusdsfutureswebsocketstreams_PublicAPIService(
 		)
 		mockClient.WebsocketStreams.WsPublic.WsCommon.Connections = []*common.WebSocketConnection{conn}
 
-		mockedJSON := `{"e":"depthUpdate","E":123456789,"T":123456788,"s":"BNBUSDT","U":157,"u":160,"pu":149,"b":[["0.0024"]],"a":[["0.0026"]],"ps":"BTCUSDT","st":1}`
+		mockedJSON := `{"e":"depthUpdate","E":123456789,"T":123456788,"s":"BNBUSDT","U":157,"u":160,"pu":149,"b":[["0.0024","10"]],"a":[["0.0026","100"]],"ps":"BTCUSDT","st":1}`
 		mockWS.QueueMessage([]byte(mockedJSON))
 
 		resp, err := mockClient.WebsocketStreams.PublicAPI.DiffBookDepthStreams().Symbol("btcusdt").Execute()
@@ -463,7 +463,7 @@ func Test_binancederivativestradingusdsfutureswebsocketstreams_PublicAPIService(
 		)
 		mockClient.WebsocketStreams.WsPublic.WsCommon.Connections = []*common.WebSocketConnection{conn}
 
-		mockedJSON := `{"e":"depthUpdate","E":123456789,"T":123456788,"s":"BNBUSDT","U":157,"u":160,"pu":149,"b":[["0.0024"]],"a":[["0.0026"]],"ps":"BTCUSDT","st":1}`
+		mockedJSON := `{"e":"depthUpdate","E":123456789,"T":123456788,"s":"BNBUSDT","U":157,"u":160,"pu":149,"b":[["0.0024","10"]],"a":[["0.0026","100"]],"ps":"BTCUSDT","st":1}`
 		mockWS.QueueMessage([]byte(mockedJSON))
 
 		resp, err := mockClient.WebsocketStreams.PublicAPI.RpiDiffBookDepthStreams().Symbol("btcusdt").Execute()